requirements.txt
Brotli==1.0.9
fonttools==4.29.1
gitdb==4.0.9
GitPython==3.1.27
Pillow==9.0.1
pypng==0.0.21
smmap==5.0.0
